Error when simple and conditional rules conflict in RuleSet
When a RuleSet provides both a simple and conditional rule for a single
syscall, in the past extrasafe would ignore the simple rule and only use
the conditional rule.

Now extrasafe will raise the same error as when two different RuleSets
provided both simple and conditional rules.

This fixes GitHub issue #6