feat: release the turbofish
Unlike functions with anonymous generic arguments (foo: impl Bar),
functions with named generic arguments can be called using the
turbofish-notation `foo::<Baz>` in the case of the compiler unable to
infer the type themself. IMHO, in library-methods, anonymous generics
should be avoided. Hoewever, they are perfectly fine for internal
methods and functions or in an initial implementation.

In memory of Anna Harren.